[refactor] Separation of functions
